Add `BroadcastableShape` trait to tpp.add
Instead of having free functions to check if the operand are
broadcastable use an operation trait.
